requirements.txt
asgiref==3.5.2
astroid==2.11.7
attrs==22.1.0
autopep8==1.7.0
black==22.10.0
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
Django==4.1.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
inflection==0.5.1
isort==5.10.1
jsonschema==4.16.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.4
pycodestyle==2.9.1
pylint==3.0.0a5
pyrsistent==0.18.1
python-snappy==0.6.1
pytz==2022.4
PyYAML==6.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
winkerberos==0.9.0
wrapt==1.14.1
zipp==3.9.0
